+++ src/main/java/com/example/finmonitor/api/dto/dashboard/BankStatDto.java
@@ -1,4 +1,4 @@
-package com.example.finmonitor.application.dto.dashboard;
+package com.example.finmonitor.api.dto.dashboard;
 
 import java.math.BigDecimal;
 import java.util.UUID;
