+++ src/main/java/com/example/finmonitor/api/dto/dashboard/TransactionTypeDto.java
@@ -1,4 +1,4 @@
-package com.example.finmonitor.application.dto.dashboard;
+package com.example.finmonitor.api.dto.dashboard;
 
 import com.example.finmonitor.application.enums.TxnType;
